fix(api): remove duplicate registry wrapper
